refactor(LottoOptionInputView): 복잡해진 Input 역할을 두 클래스에 나누어 분리
